+++ src/components/Onboarding/Registration/index.tsx
@@ -7,7 +7,7 @@ interface ButtonProps {
 export const TypeButton: React.FC<ButtonProps> = ({ isChecked, onClick, children }) => {
     return(
         <button
-        className={`min-w-[10rem] items-center justify-center py-[0.88rem] rounded-[0.31rem] body1-medium ${
+        className={`h-screen min-w-[8rem] max-h-[3.5rem] items-center justify-center rounded-[0.31rem] body1-medium ${
           isChecked ? 'bg-grayscale-80 text-dark' : 'bg-grayscale-30 text-grayscale-80'
         }`}
         onClick={onClick}
@@ -20,7 +20,7 @@ export const TypeButton: React.FC<ButtonProps> = ({ isChecked, onClick, children
 export const ConceptButton: React.FC<ButtonProps> = ({ isChecked, onClick, children }) => {
     return(
         <button
-        className={`items-center justify-center px-[3rem] py-[3rem] rounded-[0.31rem] body1-medium ${
+        className={`items-center justify-center py-[3rem] rounded-[0.31rem] body1-medium ${
           isChecked ? 'bg-grayscale-80 text-dark' : 'bg-grayscale-30 text-grayscale-80'
         }`}
         onClick={onClick}
@@ -28,4 +28,17 @@ export const ConceptButton: React.FC<ButtonProps> = ({ isChecked, onClick, child
         {children}
       </button>
     );
+};
+
+export const FeatureButton: React.FC<ButtonProps> = ({ isChecked, onClick, children }) => {
+  return(
+      <button
+      className={`h-screen items-center justify-center max-h-[9rem] py-[1rem] rounded-[0.31rem] body1-medium ${
+        isChecked ? 'bg-grayscale-80 text-dark' : 'bg-grayscale-30 text-grayscale-80'
+      }`}
+      onClick={onClick}
+      >
+      {children}
+    </button>
+  );
 };
