+++ src/pages/Onboarding/UserRegistration/Location/page.tsx
@@ -1,4 +1,4 @@
-import { TypeButton } from "@/components/Onboarding/hobbyRegister";
+import { TypeButton } from "@/components/Onboarding/Registration";
 import { Progress } from "@/components/ui/progress";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
