+++ src/pages/Onboarding/UserRegistration/Price/page.tsx
